--- src/main/java/com/yapp/artie/archive/domain/category/Category.java
@@ -1,24 +1,0 @@
-package com.yapp.artie.archive.domain.category;
-
-import com.yapp.artie.global.common.BaseEntity;
-import com.yapp.artie.user.domain.User;
-import lombok.*;
-
-import javax.persistence.*;
-
-@Entity
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Category extends BaseEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;
-
-  @Column(nullable = false)
-  private String name;
-}
