--- src/main/java/com/yapp/artie/archive/domain/exhibit/Exhibit.java
@@ -1,38 +1,0 @@
-package com.yapp.artie.archive.domain.exhibit;
-
-
-import com.yapp.artie.archive.domain.category.Category;
-import com.yapp.artie.global.common.BaseEntity;
-import com.yapp.artie.user.domain.User;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.DynamicInsert;
-
-import javax.persistence.*;
-
-@Entity
-@Table(name = "post")
-@Getter
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@DynamicInsert
-public class Exhibit extends BaseEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "user_id", nullable = false)
-  private User user;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "category_id", nullable = false)
-  private Category category;
-
-  @Embedded
-  private ExhibitContents contents;
-
-  @Embedded
-  private Publication publication;
-}
