+++ src/main/java/com/yapp/artie/domain/archive/domain/category/HomeCategoryId.java
@@ -1,0 +1,17 @@
+package com.yapp.artie.domain.archive.domain.category;
+
+import lombok.*;
+
+import javax.persistence.Embeddable;
+import java.io.Serializable;
+
+@Getter
+@Embeddable
+@EqualsAndHashCode
+@AllArgsConstructor
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class HomeCategoryId implements Serializable {
+
+  private Long categoryId;
+  private Long userId;
+}
