+++ src/main/java/com/yapp/artie/domain/archive/domain/tag/Tag.java
@@ -1,0 +1,24 @@
+package com.yapp.artie.domain.archive.domain.tag;
+
+import com.yapp.artie.domain.user.domain.User;
+import com.yapp.artie.global.common.BaseEntity;
+import lombok.*;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Tag extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "user_id", nullable = false)
+  private User user;
+
+  @Column(nullable = false)
+  private String name;
+}
