+++ src/main/java/com/yapp/artie/domain/notice/domain/Notice.java
@@ -1,0 +1,29 @@
+package com.yapp.artie.domain.notice.domain;
+
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.springframework.data.annotation.CreatedDate;
+
+import javax.persistence.*;
+import java.time.LocalDateTime;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Notice {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @Column(nullable = false)
+  private String title;
+
+  @Column(columnDefinition = "text")
+  private String contents;
+
+  @Column(nullable = false, updatable = false)
+  @CreatedDate
+  private LocalDateTime createdAt;
+}
