+++ src/main/java/com/yapp/artie/domain/question/domain/Question.java
@@ -1,0 +1,33 @@
+package com.yapp.artie.domain.question.domain;
+
+import com.yapp.artie.domain.user.domain.User;
+import com.yapp.artie.global.common.BaseEntity;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import javax.persistence.*;
+
+@Entity
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class Question extends BaseEntity {
+
+  @Id
+  @GeneratedValue(strategy = GenerationType.IDENTITY)
+  private Long id;
+
+  @ManyToOne(fetch = FetchType.LAZY)
+  @JoinColumn(name = "user_id")
+  private User user;
+
+  @Column(nullable = false)
+  private String title;
+
+  @Column(columnDefinition = "text")
+  private String contents;
+
+  private String image;
+
+  private String replyContents;
+}
