+++ src/main/java/com/yapp/artie/domain/user/dto/response/CreateUserResponseDto.java
@@ -1,0 +1,14 @@
+package com.yapp.artie.domain.user.dto.response;
+
+import io.swagger.v3.oas.annotations.media.Schema;
+
+@Schema(description = "회원 Response")
+public class CreateUserResponseDto {
+
+  @Schema(description = "아이디")
+  public Long id;
+
+  public CreateUserResponseDto(Long id) {
+    this.id = id;
+  }
+}
