--- src/main/java/com/yapp/artie/user/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package com.yapp.artie.user.repository;
-
-import com.yapp.artie.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-
-  Optional<User> findByUid(String uid);
-}
